Adds simple overloaded method for #186
At some point, I'll rewrite the CustomArgument class because I think it's a bit of a mess. I have to write a specification to figure out everything I want to do with it, but for now I think this is a good step in the right direction. With this commit, you should be able to just use the overloaded constructor with the lambda expression `new CustomArgument("nodeName", (sender, input) -> { ... }, false);`