Apply 'Essentials fix' to 1.16.5
For context as to what this commit and the previous commit fixes, basically when the CommandAPI does its reload step, the datapackResources command dispatcher's list of commands contains ALL commands, including bukkit and Minecraft ones. This conflicts with commands with the same name, so Essentials' /tp command replaces Minecraft's /tp command, converting it into /minecraft:tp, which isn't picked up by functions. If a function has /tp, then it tries to run the Essentials' one and that blows everything up. This fixes that by giving it the "correct" command context. Hopefully.