Fixes #316 to do with reloading on Spigot
We don't need to directly access the completable future in a thread that isn't the main thread because it's already being processed in the manageBlocked call.

Additionally, we could go directly to Bukkit.getServer().syncCommands() instead of accessing that via NMS, but there's no real incentive to do it one way or the other.