Replace NMS VarHandles with SafeVarHandles
This makes improvements to the compile-time type safety of using VarHandle#get. By using SafeVarHandle#get, we know that the input type will be correct and the return type will be correct. This saves unnecessary casting as well.

There is an exception to this which is helpMapTopics.get() which needs a SimpleHelpMap. Bukkit.getServer().getHelpMap() returns a HelpMap, so we have to explicitly cast this (which looks unnecessary, but is indeed "safer"!)

Also renamed  ParticleParamRedstone_f/g to particleParamRedstoneSize as it represents the size of a dust particle.