Implementation of particle tests (2/4)
Notably, we need to be able to get the Minecraft namespaced key from a particle. Spigot (nor paper) expose this with their API, so we decide to implement that with reflection directly from CraftParticle, which we assume has the same enum names as the Particle enum.
Fixes a bug where the ParticleArgument was using the wrong underlying class in the test suite. Update 1.18 tests with Minecraft versions. Starts implementing some particle argument tests with repeated tests for greater coverage.