Added a small note at the bottom of the particle argument stating how to use vibrations
This isn't the best way to implement it - I'd rather shove everything into Examples and then include Java and Kotlin code, but I'm not feeling like that right now and we've got more important things to do. If anyone's interested in doing this, you're more than welcome to do so, but I highly doubt anyone's going to really notice or care.