Removed private constructor that uses Brigadier type IntegerArgumentType
Having constructors that use types that are inaccessible to the end user/developer results in compilation errors because (all) constructors have to be resolved for compilation. Hence, we can't have constructors (even private ones!) that have types that a developer may not have - in this case, Brigadier isn't a required dependency, so most developers won't have that in their development environment and won't be able to resolve this class at compile time.