Adds a togglable button to the top of the documentation
From https://dev.to/codebubb/create-a-simple-on-off-slide-toggle-with-css-db8, converted to CSS (from SCSS) and then adjusted some of the styling to better fit the main house-style. The current "top of the document" label is offset incorrectly and I've added some bodgy code to deal with it, but I don't know if that's the best solution.