Fix runtime exception with missing reference to Adventure's Component
I honestly DO NOT know why this fix works and why this doesn't work with the original code. I've stared at bytecode for ages and haven't cracked it. I think we'll just end up spending so much time staring at this nonsense that it's not going to be worth anyone's time, even from a Java learning perspective. My current theory is it's probably a funky bug in Java's compiler or something. Who knows.