Fix `CommandAPIBukkit#postCommandRegistration` and conflicts with the minecraft namespace
This makes all the tests work as expected

When VanillaCommandWrappers are executed, they defer to the Brigadier dispatcher using their name. This was causing problems with permissions and conflicts when the minecraft namespace was involved. In certain cases, CommandMap commands were referencing different commands in the Brigadier dispatcher.

The changes here fix all that so CommandMap commands are always referencing the correct node in the Brigadier dispatcher.