Initial commit of NMS for 1.20.5. Adds font reference for logo
The NMS_1_20_R4 still has a lot of errors. We're still linking against paper 1.20.4 until a paper 1.20.5 library becomes available. Notably the thing that needs extra care and attention to are loot tables and the way that we convert NMS objects to component-style strings. We need to investigate how to serialize JSON correctly - a new parameter was introduced that I am unfamiliar with.