Implement `DynamicMultiLiteralArgument`
Acts like `MultiLiteralArgument`, but literals can be changed and differ for each CommandSender

Resolves https://github.com/JorelAli/CommandAPI/issues/513

Uses Paper's `AsyncPlayerSendCommandsEvent` and Velocity's `PlayerAvailableCommandsEvent` to change the client's view of the command tree. Note that a similar event does not exist on Spigot, so the suggestions are not dynamic, though the parsing still is.

`DifferentClientNode` class added to handle creating client-server command tree de-syncs (I believe this can be used to make `FlagsArgument` platform-agnostic)