Refactor parseArgument() method in ListArgumentCommon to be simpler
We changed the hashmap to String -> T. This saves requiring an additional loop to do effectively a hashmap lookup. Refactor the code to use fast-fails which makes adding to the list neater and makes errors easier to identify and extend. Remove cursor tracking in favour of shoving everything into the context instead.