Create `commandapi-bukkit-test-toolkit`
Create `automated-tests-shaded` example project to test using the test toolkit

Resolves https://github.com/JorelAli/CommandAPI/issues/356

Alternative idea to current work on `dev/public-test-suite`

Attempts to allow developers to test their commands independently of any version or nms stuff. Currently, most of the methods throw an UnimplementedMethodException, but I have implemented enough to get the simple `PingPongCommandTests` file to work as a proof of concept.

TODO:
- Implement the methods required to test:
  - `commandapi-bukkit-plugin` based projects
  - Common arguments
  - Suggestions
- Write documentation
- Maybe add a module to more rigorously test the toolkit without cluttering the example project

I think if we can get a simple version working and published as a snapshot, users can try it out and identify methods that need to be implemented.