Implement ArgumentType suggestions for test toolkit
I think the way I set up the Parser builder to support suggestions makes sense. It works at least.

Parser now returns Result object, which tracks whether an object was returned or an exception was thrown, as well as a SuggestionProvider for the current location. Default methods added to Parser to handle this result appropriately when either parsing or suggesting.

Moved Parser into its own package and extracted some of the inner classes and interfaces. Package-protected access accomplishes the same sort of encapsulation.

Added methods to CommandAPITestUtilities to assert that suggestions are located at a specific offset index.

Added tests~