Update shade plugin to latest version
When running a parallel build (maven's -T option), the process would sometimes seize up and stop making progress. The modules currently being processed all seemed to get stuck when executing a shade goal.

After updating the shade plugin and running the build for a while on infinite loop, it didn't get stuck again. Perhaps there was a bug in the shade plugin that was fixed somewhere along v3.3.0 to v3.6.0. Hard to tell though with random issues.